requirements.txt
streamlit
mysql-connector-python
pandas
numpy
matplotlib
